Add set token to blacklist method and refactoring logout endpoint

+++ backend/auth/services.py
@@ -1,20 +1,30 @@
 from fastapi import Request, Response
 
+from time import time
 from datetime import datetime, timezone, timedelta
 
 from src.settings import jwt_settings
 from users.repositories import UsersRepository
 from core.utils.password import hashing_password, verify_password
-from core.utils.jwt import create_jwt_token
+from core.utils.jwt import create_jwt_token, verify_jwt_token
 from core.repositories.redis_base import RedisBaseRepository
 
 from .schemas import AccessTokenResponseSchema, UserRegistrationSchema, UserLoginSchema
-from .exceptions import EmailAlreadyRegistered, EmailOrPasswordIncorrect
+from .exceptions import EmailAlreadyRegistered, EmailOrPasswordIncorrect, TokenMissing
 
 
 class BlacklistTokensService:
     def __init__(self, redis_repository: RedisBaseRepository):
         self.redis_repository = redis_repository
+    
+    async def set_token_to_blacklist(self, payload: dict) -> None:
+        exp = payload.get('exp')
+        jti = payload.get('jti')
+
+        current_time = int(time())
+        ttl = exp - current_time
+
+        await self.redis_repository.setex(f'blacklist:{jti}', 1, ttl)
 
 
 class JWTTokensService:
@@ -83,6 +93,15 @@ class AuthService:
         return AccessTokenResponseSchema(access_token=access_token)
 
     async def logout(self, request: Request, response: Response) -> dict[str, str]:
+        refresh_token = request.cookies.get('refresh_token')
+
+        if refresh_token is None:
+            raise TokenMissing()
+        
+        payload = verify_jwt_token(refresh_token)
+
+        await self.blacklist_tokens_service.set_token_to_blacklist(payload)
+
         response.delete_cookie('refresh_token')
 
         return {'message': 'User successfully logged out'}
